remove older drag and drop code
