+++ met-web/src/components/common/Dragndrop/index.tsx
@@ -1,98 +1,46 @@
-import React, { useRef } from 'react';
+import { Draggable, Droppable } from '@hello-pangea/dnd';
 import { Box } from '@mui/material';
-import type { Identifier, XYCoord } from 'dnd-core';
-import { useDrag, useDrop } from 'react-dnd';
+import React from 'react';
 
-export interface DragItemProps {
-    index: number;
-    testId?: string;
-    moveItem: (dragIndex: number, hoverIndex: number) => void;
-    name: string;
+interface MetDroppableProps {
     children: React.ReactNode;
+    droppableId: string;
 }
-
-interface Draggable {
-    index: number;
-    id: string;
-    type: string;
-}
-
-const DragItem = ({ index, moveItem, testId, name, children }: DragItemProps) => {
-    const ref = useRef<HTMLDivElement>(null);
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [{ handlerId }, drop] = useDrop<Draggable, void, { handlerId: Identifier | null }>({
-        accept: name,
-        collect(monitor) {
-            return {
-                handlerId: monitor.getHandlerId(),
-            };
-        },
-        hover(item: Draggable, monitor) {
-            if (!ref.current) {
-                return;
-            }
-            const dragIndex = item.index;
-            const hoverIndex = index;
-
-            // Don't replace items with themselves
-            if (dragIndex === hoverIndex) {
-                return;
-            }
-
-            // Determine rectangle on screen
-            const hoverBoundingRect = ref.current?.getBoundingClientRect();
-
-            // Get vertical middle
-            const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-
-            // Determine mouse position
-            const clientOffset = monitor.getClientOffset();
-
-            // Get pixels to the top
-            const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
-
-            // Only perform the move when the mouse has crossed half of the items height
-            // When dragging downwards, only move when the cursor is below 50%
-            // When dragging upwards, only move when the cursor is above 50%
-
-            // Dragging downwards
-            if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-                return;
-            }
-
-            // Dragging upwards
-            if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-                return;
-            }
-
-            // Time to actually perform the action
-            moveItem(dragIndex, hoverIndex);
-
-            // Note: we're mutating the monitor item here!
-            // Generally it's better to avoid mutations,
-            // but it's good here for the sake of performance
-            // to avoid expensive index searches.
-            item.index = hoverIndex;
-        },
-    });
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [{ isDragging }, drag] = useDrag({
-        type: name,
-        item: () => {
-            return { testId, index };
-        },
-
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        collect: (monitor: any) => ({
-            isDragging: monitor.isDragging(),
-        }),
-    });
-
-    drag(drop(ref));
-
-    return <Box ref={ref}>{children}</Box>;
+export const MetDroppable = ({ droppableId, children }: MetDroppableProps) => {
+    return (
+        <Droppable droppableId={droppableId}>
+            {(provided: any) => (
+                <Box {...provided.droppableProps} ref={provided.innerRef}>
+                    {children}
+                    {provided.placeholder}
+                </Box>
+            )}
+        </Droppable>
+    );
 };
 
-export { DragItem };
+interface MetDraggableProps {
+    index: number;
+    children: React.ReactNode;
+    draggableId: string;
+    marginBottom?: string | number;
+}
+export const MetDraggable = ({ children, draggableId, index, marginBottom }: MetDraggableProps) => {
+    return (
+        <Draggable key={draggableId} draggableId={draggableId} index={index}>
+            {(provided: any) => (
+                <Box
+                    ref={provided.innerRef}
+                    {...provided.draggableProps}
+                    {...provided.dragHandleProps}
+                    sx={{
+                        ...provided.draggableProps.style,
+                        marginBottom: marginBottom || '1em',
+                    }}
+                >
+                    {children}
+                </Box>
+            )}
+        </Draggable>
+    );
+};
