+++ met-web/src/index.tsx
@@ -8,8 +8,6 @@ import { BaseTheme } from 'styles/Theme';
 import { Formio } from '@formio/react';
 import MetFormioComponents from 'met-formio';
 import '@bcgov/bc-sans/css/BCSans.css';
-import { DndProvider } from 'react-dnd';
-import { HTML5Backend } from 'react-dnd-html5-backend';
 
 Formio.use(MetFormioComponents);
 Formio.Utils.Evaluator.noeval = false;
@@ -19,13 +17,11 @@ const root = ReactDOM.createRoot(document.getElementById('root')!);
 root.render(
     // <React.StrictMode>
     <Provider store={store}>
-        <DndProvider backend={HTML5Backend}>
-            <ThemeProvider theme={BaseTheme}>
-                <StyledEngineProvider injectFirst>
-                    <App />
-                </StyledEngineProvider>
-            </ThemeProvider>
-        </DndProvider>
+        <ThemeProvider theme={BaseTheme}>
+            <StyledEngineProvider injectFirst>
+                <App />
+            </StyledEngineProvider>
+        </ThemeProvider>
     </Provider>,
     // </React.StrictMode>
 );
